+++ .changeset/smooth-olives-cheat.md
@@ -2,4 +2,4 @@
 '@ag.ds-next/details': major
 ---
 
-Create Detail component
+Create Details component
