+++ docs/components/designSystemComponents.tsx
@@ -148,4 +148,4 @@ export {
 } from '@ag.ds-next/skeleton';
 export { Autocomplete } from '@ag.ds-next/autocomplete';
 export { Combobox } from '@ag.ds-next/combobox';
-export { Detail } from '@ag.ds-next/details';
+export { Details } from '@ag.ds-next/details';
