+++ packages/details/docs/code.mdx
@@ -1,3 +1,3 @@
 ## Props
 
-<ComponentPropsTable name="Detail" />
+<ComponentPropsTable name="Details" />
