+++ packages/details/docs/overview.mdx
@@ -1,5 +1,5 @@
 ---
-title: Detail
+title: Details
 description: A disclosure widget for supplimental information that explains a component, section or flow.
 group: Layout
 storybookPath: /story/layout-details--basic
@@ -8,10 +8,10 @@ storybookPath: /story/layout-details--basic
 Information is visible only when the widget is toggled into an "open" state. Unlike Accordion, inner content is surfacable in a page search (cmd/ctrl-f).
 
 ```jsx live
-<Detail label="Details">
+<Details label="Details">
 	<Text as="p">
 		This is a small paragraph of text that is supplimentary to the main page
 		content.
 	</Text>
-</Detail>
+</Details>
 ```
